WIP on groups.
We'll need intersection buffers for this to work in all cases.
Since a group can return an arbitrary number of intersections, we need
to make the returned number of intersections dynamic. This will also be
a cleaner solution than setting the intersection count to 4 like we do
now (because cones can have 4 intersections). We can still preallocate
this memory because we can still calculate the max number of
intersections based on the amount (and type) of objects in the world.